Update docs and tests for permission and policy management
